tests: requests testability
